requirements.txt
Flask==2.2.5
tensorflow-cpu==2.10.0
keras==2.10.0
numpy==1.23.5
opencv-python-headless
Pillow
reportlab
werkzeug
gunicorn

